// workspaces/apps/packages/goldstack-home/src/components/DownloadInstructions.d.ts
import type { DocLink } from '@goldstack/goldstack-api/dist/src/utils/docLinks';
export declare const DownloadInstructions: (props: {
  projectId: string;
  packageId: string;
  downloadUrl: string;
  docLinks: DocLink[] | undefined;
}) => React.ReactNode;
//# sourceMappingURL=DownloadInstructions.d.ts.map